backend/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
aiosqlite==0.21.0
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.30.0
attrs==23.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.4.0
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
chefshatgym==2.4.0.4
click==8.1.7
cloud-init==24.4
cloudpickle==3.1.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.2
cryptography==3.4.8
cycler==0.12.1
dataclasses-json==0.6.7
dbus-python==1.2.18
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.40.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.71.0
gym==0.26.2
gym-notices==0.0.8
gyp==0.1
h11==0.14.0
h5py==3.13.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.4.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==2.14.0
keyring==23.5.0
kiwisolver==1.4.8
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
llama-cloud==0.0.11
llama-index==0.10.58
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.58
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.27
llama-index-multi-modal-llms-openai==0.1.8
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.32
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.8.0
mdurl==0.1.2
ml-dtypes==0.2.0
more-itertools==8.10.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
nltk==3.8.1
numpy==1.26.1
oauthlib==3.2.0
ollama==0.3.1
openai==1.37.1
opencv-python==4.8.1.78
opt_einsum==3.4.0
packaging==24.1
pandas==2.1.1
passlib==1.7.4
pillow==10.2.0
protobuf==4.25.7
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
pydantic==2.8.2
pydantic-settings==2.9.1
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymediawiki==0.7.4
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypdf==4.1.0
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.1
rsa==4.9.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlmodel==0.0.24
starlette==0.46.2
streamlit==1.37.0
striprtf==0.0.26
systemd-python==234
tenacity==8.5.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tiktoken==0.7.0
toml==0.10.2
toolz==0.12.1
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.4.1
tqdm==4.66.4
trash==0.1.0
Twisted==22.1.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2024.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
uvicorn==0.34.2
wadllib==1.3.6
waitress==3.0.0
watchdog==4.0.1
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.4
zipp==1.0.0
zope.interface==5.4.0
